feat: implement batch interface
