+++ test/config/test.config.toml
@@ -13,7 +13,7 @@ SequencesTxFileName = "sequencesender.json"
 GasOffset = 80000
 WaitPeriodPurgeTxFile = "60m"
 MaxPendingTx = 1
-SanityCheckRPCURL = "http://127.0.0.1:8123"
+RPCURL = "http://127.0.0.1:8123"
 	[SequenceSender.StreamClient]
 		Server = "127.0.0.1:6900"
 	[SequenceSender.EthTxManager]
