+++ test/config/test.kurtosis_template.toml
@@ -27,7 +27,7 @@ SequencesTxFileName = "sequencesender.json"
 GasOffset = 80000
 WaitPeriodPurgeTxFile = "15m"
 MaxPendingTx = 1
-SanityCheckRPCURL = "http://127.0.0.1:8123"
+RPCURL = "http://127.0.0.1:8123"
 	[SequenceSender.StreamClient]
 		Server = "127.0.0.1:${zkevm_data_streamer_port}"
 	[SequenceSender.EthTxManager]
